trello_backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-jwt>=1.11,<2.0
psycopg2>=2.9,<3.0
channels>=3.0,<4.0
channels_redis>=3.2,<4.0
Pillow>=8.0,<10.0
